# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Todo": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"completed": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"title",
					"completed"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "BunHono",
		"contact": {}
	},
	"paths": {
		"/todos": {
			"get": {
				"operationId": "GetTodos",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Todo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all todos",
				"tags": [
					"Todos"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "AddTodo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"todo": {
											"$ref": "#/components/schemas/Todo"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"todo",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Add a new todo",
				"tags": [
					"Todos"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The todo item to add",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"completed": {
										"type": "boolean"
									},
									"title": {
										"type": "string"
									}
								},
								"required": [
									"completed",
									"title"
								],
								"type": "object",
								"description": "The todo item to add"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}